Replace catalogue with the sql_catalogue.
